+++ WebApiTerra1000/Controllers/ContragentController.cs
@@ -9,7 +9,8 @@ using System;
 using System.Net;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using WebApiTerra1000.Common;
+using WebApiTerra1000.Utils;
+using static WebApiTerra1000.Utils.TerraEnums;
 
 namespace WebApiTerra1000.Controllers
 {
@@ -19,6 +20,7 @@ namespace WebApiTerra1000.Controllers
 
         public ContragentController(ILogger<ContragentController> logger, ISessionFactory sessionFactory) : base(logger, sessionFactory)
         {
+            //
         }
 
         #endregion
@@ -28,12 +30,26 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/contragents-test/")]
-        public ContentResult GetContragentsTest()
+        public ContentResult GetContragentsTest(FormatType format = FormatType.Raw)
         {
+            //return TaskHelper.RunTask(new Task<ContentResult>(() =>
+            //{
+            //    string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetException);
+            //    SqlSimpleEntity simple = new(response);
+            //    return format switch
+            //    {
+            //        FormatType.Json => TerraUtils.GetResult(FormatType.Json, simple.SerializeAsJson(), HttpStatusCode.OK),
+            //        FormatType.Xml => TerraUtils.GetResult(FormatType.Xml, simple.SerializeAsXml(), HttpStatusCode.OK),
+            //        FormatType.Html => TerraUtils.GetResult(FormatType.Html, simple.SerializeAsHtml(), HttpStatusCode.OK),
+            //        FormatType.Text => TerraUtils.GetResult(FormatType.Text, simple.SerializeAsText(), HttpStatusCode.OK),
+            //        FormatType.Raw => TerraUtils.GetResult(FormatType.Text, simple.SerializeAsText(), HttpStatusCode.OK),
+            //        _ => throw TerraUtils.GetArgumentException(nameof(format)),
+            //    };
+            //}), format);
             return TaskHelper.RunTask(new Task<ContentResult>(() =>
             {
                 XDocument response = new(
-                    new XElement("Response",
+                    new XElement(TerraConsts.Response,
                         new XElement("Contragents",
                             new XElement("Result") { Value = "Success" }
                         )
@@ -44,13 +60,13 @@ namespace WebApiTerra1000.Controllers
                     StatusCode = (int)HttpStatusCode.OK,
                     Content = response.ToString()
                 };
-            }));
+            }), format);
         }
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/contragent/")]
-        public ContentResult GetContragent(int id)
+        public ContentResult GetContragent(int id, FormatType format = FormatType.Raw)
         {
             return TaskHelper.RunTask(new Task<ContentResult>(() =>
             {
@@ -68,13 +84,14 @@ namespace WebApiTerra1000.Controllers
                     StatusCode = (int)HttpStatusCode.OK,
                     Content = doc.ToString()
                 };
-            }));
+            }), format);
         }
 
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/contragents/")]
-        public ContentResult GetContragents(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10)
+        public ContentResult GetContragents(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10, 
+            FormatType format = FormatType.Raw)
         {
             return TaskHelper.RunTask(new Task<ContentResult>(() =>
             {
@@ -95,7 +112,7 @@ namespace WebApiTerra1000.Controllers
                     StatusCode = (int)HttpStatusCode.OK,
                     Content = doc.ToString()
                 };
-            }));
+            }), format);
         }
 
         #endregion