+++ WebApiTerra1000/Utils/TerraConsts.cs
@@ -1,0 +1,22 @@
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+namespace WebApiTerra1000.Utils
+{
+    public static class TerraConsts
+    {
+        public const string Response = @"Response";
+        public const string Info = @"Info";
+        public const string Error = @"Error";
+        public const string Message = @"Message";
+        public const string Value = @"Value";
+        public const string Description = @"Description";
+        public const string FilePath = @"FilePath";
+        public const string LineNumber = @"LineNumber";
+        public const string MemberName = @"MemberName";
+        public const string Exception = @"Exception";
+        public const string InnerException = @"InnerException";
+        public const string Simple = @"Simple";
+        public const string Items = @"Items";
+    }
+}