+++ WebApiTerra1000/Utils/TerraEnums.cs
@@ -1,9 +1,9 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-namespace Terra.Common
+namespace WebApiTerra1000.Utils
 {
-    public static class Enums
+    public static class TerraEnums
     {
         public enum FormatType
         {
@@ -11,7 +11,7 @@ namespace Terra.Common
             Xml,
             Html,
             Text,
-            Auto
+            Raw
         }
     }
 }