+++ WebApiTerra2Tests/Properties/AssemblyInfo.cs
@@ -31,5 +31,5 @@
 //
 // You can specify all the values or you can default the Revision and Build Numbers
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("0.5.620.0")]
-[assembly: AssemblyFileVersion("0.5.620.0")]
+[assembly: AssemblyVersion("0.5.630.0")]
+[assembly: AssemblyFileVersion("0.5.630.0")]