+++ CHANGELOG.md
@@ -17,10 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redmine 1600. Подготовить шаблон инструкции "Действия при ошибках печати"
 - Redmine 1608. DbScales. Таблица Labels. Удалить поле Label
 
-## [0.5.620] - 2021-09-20
+## [0.5.630] - 2021-09-21
 ### Fixed
-- DbDwh v.0.5.620
-- WebApiTerra1000 v.0.5.620
+- DbDwh v.0.5.630
+- WebApiTerra1000 v.0.5.630
 
 ## [0.5.610] - 2021-09-17
 ### Fixed