+++ DbDwh/CHANGELOG.md
@@ -6,13 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.620] - 2021-09-20
+## [0.1.630] - 2021-09-21
 ### Changed
 - IIS\Functions\GetRefShipmentsByDocDate
+- DBO\Functions\fnGetDeprecated
 ### Added
 - IIS\Functions\fnGetDeliveryPlaces
 - DBO\Functions\fnCheckDates
 - DBO\Functions\fnCheckRowCount
+- DBO\Functions\fnGetXmlSimpleV1
+- DBO\Functions\fnGetXmlSimpleV2
 
 ## [0.1.600] - 2021-09-15
 ### Changed